Price changed to Double from Int

+++ Abstract_Class_Task/Models/Product.cs
@@ -9,7 +9,7 @@ namespace Abstract_Class_Task.Models
     internal class Product
     {
         public string Name { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
         public int Count { get; set; }
 
         public void Detail()
@@ -21,8 +21,8 @@ namespace Abstract_Class_Task.Models
         public void Discount()
         {
             Console.WriteLine("Endirim %-ni daxil edin:");
-            int.TryParse(Console.ReadLine(), out int discount);
-            int newprice = Price - (Price * discount / 100);
+            double.TryParse(Console.ReadLine(), out double discount);
+            double newprice = Price - (Price * discount / 100);
             Console.WriteLine($"Endirimli giymet: {newprice}");
         }
